pyproject.toml
[project]
name = "arca-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "firebase-admin",
    "firebase-functions>=0.4.3",
    "google-genai>=1.45.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "natal>=0.9.6",
    "numpy>=2.3.4",
    "posthog>=6.7.8",
    "pydantic>=2.12.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
]

[dependency-groups]
dev = [
    "firebase-admin>=7.1.0",
    "ipykernel>=7.0.1",
    "mypy>=1.18.2",
    "pandas>=2.3.3",
    "pandas-stubs>=2.3.2.250926",
    "pyarrow>=22.0.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.3.0",
    "tqdm>=4.67.1",
]
